# storage/api-docs/vendor-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medishop",
        "description": "Medishop Project",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://api.medishop.dworklabs.com/api/v1",
            "description": "Live API Server"
        },
        {
            "url": "http://192.168.1.68:8000/api/v1",
            "description": "Localhost API Server 1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Localhost API Server 2"
        },
        {
            "url": "http://192.168.100.18:8000/api/v1",
            "description": "Localhost API Server 3"
        }
    ],
    "paths": {
...
        "/vendor/orders": {
            "get": {
                "tags": [
                    "Vendor Orders"
                ],
                "summary": "Get list of assigned orders for the logged-in vendor",
                "operationId": "747e630009c98afa4c338e4108412efb",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by customer name, email, mobile, or address",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John Doe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of assigned orders retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of Assign Orders."
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 12
                                                            },
                                                            "order_code": {
                                                                "type": "string",
                                                                "example": "ORD-2025-001"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 2500.75
                                                            },
                                                            "user_name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "user_email": {
                                                                "type": "string",
                                                                "example": "john@example.com"
                                                            },
                                                            "order_items_count": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "pending"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-10-20 12:30:00"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total_items": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/vendor/get-variant-batch-numbers/{id}": {
            "get": {
                "tags": [
                    "Vendor Orders"
                ],
                "summary": "Get list of batch numbers of a product variant",
                "operationId": "e7a8580ded2cd6dcf8062035c026ebe4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Variant ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of available variants batch.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "List of available variants batch."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "batch_number_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "batch_number": {
                                                        "type": "string",
                                                        "example": "776033374"
                                                    },
                                                    "stock_left": {
                                                        "type": "integer",
                                                        "example": 129
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}